+++ odoo/local-src/connector_esb/models/customer/webservice.py
@@ -101,3 +101,28 @@ class StatisticsFormWebserviceMessage(Component):
 
     def get_message(self, options):
         return self._produce_xml(self._data_for_message(options))
+
+
+class CustomerDeliveryFeeWebserviceMessage(Component):
+
+    _name = 'esb.webservice.message.customer.delivery.fee'
+    _inherit = ['esb.webservice.message.base']
+    _apply_on = ['res.partner']
+    _usage = 'ws.message.customer.delivery.fee'
+
+    def get_message(self, customer_ref):
+        data = []
+        by_pass = False
+
+        current_orders = self.env['sale.order'].search([
+            ('state', 'in', ['sent', 'sale']),
+            ('partner_id.ref', '=', customer_ref)])
+        total_amount = sum(current_orders.mapped('amount_total'))
+        if total_amount > 125:
+            by_pass = True
+        values = {
+                'totalOrderAmount': '{0:.2f}'.format(total_amount),
+                'byPassTestAmount': by_pass
+                }
+        data.append(values)
+        return self._produce_xml(data, list_item_el='result')